+++ packages/components/src/form/ImagePreviewInput/ImagePreviewInput.js
@@ -15,6 +15,8 @@ export const IMAGE_PREVIEW_INPUT_DEFAULT_PROPS = {
     uploadButton: '',
   },
   previewURL: '',
+  width: 224,
+  height: 132,
 };
 export const IMAGE_PREVIEW_INPUT_PROP_TYPES = {
   labels: PropTypes.shape({
@@ -30,6 +32,9 @@ export const IMAGE_PREVIEW_INPUT_PROP_TYPES = {
   disabled: PropTypes.bool,
   useAria: PropTypes.bool,
   noPicker: PropTypes.bool,
+  objectFit: PropTypes.string,
+  width: PropTypes.number,
+  height: PropTypes.number,
 };
 
 const ImagePreviewInput = ({
@@ -43,6 +48,9 @@ const ImagePreviewInput = ({
   disabled,
   useAria,
   noPicker,
+  objectFit,
+  width,
+  height,
 }) => {
   const [imagePreview, setImagePreview] = useState(previewURL);
   const [imageValue, setImageValue] = useState(value);
@@ -108,14 +116,15 @@ const ImagePreviewInput = ({
         <Stack spacing={2} fullWidth alignItems="flex-end">
           <ImageLoader
             src={imagePreview}
-            height={132}
-            width={224}
+            height={height}
+            width={width}
             noFlex
             radius={4}
             style={{ ...previewStyle }}
             useAria={useAria}
             bordered
             forceImage
+            objectFit={objectFit}
           />
           {!readonly && !disabled ? (
             <Box noFlex>