+++ packages/components/src/form/InputError/InputError.styles.js
@@ -1,13 +1,17 @@
 import { createStyles } from '@mantine/styles';
 import { getErrorStyle } from '../mixins/fieldStyles.mixins';
 
-export const InputErrorStyles = createStyles((theme, {}) => {
-  return {
-    error: {
-      ...getErrorStyle(theme),
-    },
-    errorIcon: {
-      color: theme.colors.fatic01,
-    },
-  };
-});
+const InputErrorStyles = createStyles((theme) => ({
+  root: {
+    flexWrap: 'nowrap',
+    alignItems: 'start',
+  },
+  error: {
+    ...getErrorStyle(theme),
+  },
+  errorIcon: {
+    color: theme.colors.fatic01,
+  },
+}));
+
+export { InputErrorStyles };