+++ packages/components/src/informative/index.js
@@ -1,7 +1,8 @@
 export * from './ActivityAnswersBar';
 export * from './ActivityCountdown';
 export * from './ActivityAccordion';
-export * from './UserDisplayItem';
+export * from './Chip';
+export * from './ChipsContainer';
 export * from './FileItemDisplay';
 export * from './PaginatedList';
 export * from './SortableList';
@@ -14,6 +15,7 @@ export * from './Calification';
 export * from './HorizontalTimeline';
 export * from './ScoresBar';
 export * from './UserCards';
+export * from './UserDisplayItem';
 export * from './UserDisplayItemList';
 export * from './ChatMessage';
 export * from './ProgressColorBar';