+++ packages/extras/CHANGELOG.md
@@ -1,5 +1,95 @@
 # @bubbles-ui/extras
 
+## 1.2.180
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.180
+  - @bubbles-ui/icons@1.2.180
+
+## 1.2.179
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.179
+  - @bubbles-ui/icons@1.2.179
+
+## 1.2.178
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.178
+  - @bubbles-ui/icons@1.2.178
+
+## 1.2.177
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.177
+  - @bubbles-ui/icons@1.2.177
+
+## 1.2.176
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.176
+  - @bubbles-ui/icons@1.2.176
+
+## 1.2.175
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.175
+  - @bubbles-ui/icons@1.2.175
+
+## 1.2.174
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.174
+  - @bubbles-ui/icons@1.2.174
+
+## 1.2.173
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.173
+  - @bubbles-ui/icons@1.2.173
+
+## 1.2.172
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.172
+  - @bubbles-ui/icons@1.2.172
+
+## 1.2.171
+
+### Patch Changes
+
+- Versions bump
+- Updated dependencies
+  - @bubbles-ui/components@1.2.171
+  - @bubbles-ui/icons@1.2.171
+
 ## 1.2.170
 
 ### Patch Changes